chore(users): move to friends controller
