+++ src/feature/api/users/users.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { UsersUseCases } from './application/use-case/users.use-case';
 import { UserRepository } from './domain/user.repository';
 import { UserRepositoryImpl } from './infrastructure/user.repository.impl';
-import { UsersController } from './users.controller';
+import { UsersController } from './presentation/controller/users.controller';
 import { UsersService } from './users.service';
 
 @Module({
